+++ rui-runtime/src/commonTest/kotlin/hu/simplexion/rui/runtime/test/manual/ForLoopTest.kt
@@ -40,9 +40,9 @@ class ForLoop(
 
     var v0 = 1
 
-    var ruiDirty0 = 0
+    var ruiDirty0 = 0L
 
-    fun ruiInvalidate0(mask: Int) {
+    fun ruiInvalidate0(mask: Long) {
         ruiDirty0 = ruiDirty0 or mask
     }
 
@@ -50,7 +50,7 @@ class ForLoop(
         if (scopeMask and 1 != 0L) return 0L
 
         it as RuiT1
-        if (ruiDirty0 and 1 != 0) {
+        if (ruiDirty0 and 1L != 0L) {
             it.p0 = v0
             it.ruiInvalidate0(1)
         }
@@ -61,7 +61,7 @@ class ForLoop(
     override fun ruiPatch(scopeMask: Long) {
         val extendedScopeMask = ruiFragment.ruiExternalPatch(ruiFragment, scopeMask)
         if (extendedScopeMask != 0L) ruiFragment.ruiPatch(extendedScopeMask)
-        ruiDirty0 = 0
+        ruiDirty0 = 0L
     }
 
     fun ruiIterator0() = IntRange(0, 10).iterator()
